+++ src/widget/multiselect/search/multiselect.search.pane.js
@@ -65,7 +65,7 @@ BI.MultiSelectSearchPane = BI.inherit(BI.Widget, {
 
     setKeyword: function (keyword) {
         var btn, o = this.options;
-        var isVisible = this.loader.getAllButtons().length > 0 && (btn = this.loader.getAllButtons()[0]) && (keyword === (o.valueFormatter(btn.getValue()) || btn.getValue()));
+        var isVisible = this.loader.getAllButtons().length > 0 && (btn = this.loader.getAllButtons()[0]) && keyword === btn.getValue();
         if (isVisible !== this.tooltipClick.isVisible()) {
             this.tooltipClick.setVisible(isVisible);
             this.resizer.attr("items")[0].height = (isVisible ? this.constants.height : 0);