+++ changelog.md
@@ -1,5 +1,10 @@
 # 更新日志
 2.0(2020-07)
+- 修复了日期类型控件设置一个不在minDate和maxDate之间的日期值时，面板灰化与翻页按钮状态不对的问题
+- BI.OB的on方法返回一个解除监听的函数
+- 修复了grid_view执行_unMount时不调用子组件的_unMount的问题
+- combo新增belowMouse属性,允许popup在点击处弹出
+- combo新增hideWhenAnotherComboOpen属性,开启则其他combo下拉时当前combo收起
 - 修复了datePicker在setValue的时候没有动态刷新可用月份的问题
 - 同步复选下拉及其面板新增getAllValue获取所有已选值
 - 同步复选下拉树及其面板新增getAllValue获取完整的选中树节点